requirements.txt
streamlit
langchain-google-genai
langchain
python-dotenv
